Add concept IRI, decrease extra work, adapt for multiple workers
Use concept label instead of simply using the last part of the
concept RDF IRI when fetching related concepts for search.

JSON returned from the search API is changed so that both the concept
label and the concept RDF IRI is returned, in addition to the
similarity score.

Users of the search function can opt out of retrieving related
concepts and retrieving dataset metadata. This is implemented
for the search API. The search CLI will automatically opt out
of retrieving data that won't be shown.

Multiple workers present a problem when the initial request goes
to a different worker than the search request, since the workers
don't share the CSRF session storage. Since none of the forms
require any form of authentication, it is safe to simply disable
CSRF (adversaries could simply make the request themselves).

Reformat some long lines in the OpenDataSemanticFramework class.