Ensure similarity graphs are up-to-date
By loading all ODSF at the start, we have a working version
available to us, even if an update should happen. Furthermore,
we now look at the concept dataset similarity matrices to see
if they are up-to-date, allowing them to be updated and not just
update in response to updated Configuration.