Use trigrams and bigrams when querying WordNet
Before, only 66,92% of WordNet synsets could be found, due to
using only a single word at a time. Now, using trigrams, bigrams
and single words, 99,23% of WordNet synsets can be matched.

The algorithm creates ngrams, then tries them out in WordNet. Any
words that matched with a synset are removed, and the remaining
words are attempted matched with a lower order of ngram.

Instead of removing the matched words completely, they are
replaced with a dummy word, so that "false" twograms aren't created.
For example, "public unidentified flying object transport" matches
with the trigram "unidentified flying object". If we just removed the
words, we would be left with "public transport", which would match with
another WordNet synset. However, this is not what the user wrote. Instead,
after finding the UFO, the text is like "public None None None transport",
and no twograms exist which can be looked up in WordNet.

Part of speech tags are not used when matching ngrams, since the
part of speech for the ngram may not always represent the part of speech
identified for the individual words. Plus, it is more important to
distinguish senses when the number of possible synsets is high. I don't
think there are cases with many competing synsets when searching with
multiple words.

This change applies to both the query and concept labels, so that they
can be matched with eachother consistenly.

One assumption made by the system is that the longest matching n-gram
is the one that should be used. Inside of the same n, all n-grams are
allowed to match, even if they overlap. The idea is that higher n
likely means more specialized words. This also reduces the number of
synsets, compared to using all words individually. Hopefully, this
helps outweight the additional processing needed to query WordNet
with all possible twograms and trigrams.