Keep stopwords, simplify chunk parser
The idea is to let some phrases, like "point of sale", remain
intact when querying WordNet. "of" would be removed since it's
a stopword, which would leave "point sale" as a word to query in
WordNet. In general, when we are trying to match words in a
semantic fashion, we need all the help we can get, even if they're
stopwords. Removing stopwords is still important for keyword-based
traditional information retrieval, like TF-IDF, where they would
just generate noise and reduce the ability to discriminate.